feat: move serde serializers to stacks_common
